Add functions for parsing the iovctl config file
Add two functions for parsing the iovctl config file.  The config
file is parsed using libucl[1], which accepts most YAML files and
a superset of JSON.  The first function is an ad-hoc parser that
searches the file for the PF.DEVICE configuration value.  We need
to know that value in order to fetch the schema from the kernel,
and we need the schema in order to be able to fully parse the file.

The second function parses the config file and validates it
against a schema.  This function will exit with an error message
if any validation error occurs.  If it succeeds, the configuration
is returned as an nvlist suitable for passing to the kernel.

[1] https://github.com/vstakhov/libucl

Differential Revision:	https://reviews.freebsd.org/D86
Reviewed by:		jhb
MFC after: 		1 month
Sponsored by:		Sandvine Inc.


git-svn-id: https://svn.freebsd.org/base/head@279460 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f