Fix a long-standing bug with the early MAC address initialisation path, which showed up after I started changing addresses this early.
It turns out that there's some other malarky going on behind the scenes
in the HAL and merely setting the net80211/ifp mac address this early
isn't enough.  If the MAC is set from kenv at attach time, the HAL
also needs to be programmed early.

Without this, the VAP wouldn't work enough for finishing association -
probe requests would be fine as they're broadcast, but association
request would fail.


git-svn-id: https://svn.freebsd.org/base/head@280802 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f