Revert	r280449; Permit multiple arguments for the nonnull attribute.
For the benefit of anyone that may be struggling to port
FreeBSD to gcc 2.8 (or older) avoid using variadic macros.

MFC	r280700 (partial);

Bring new attribute:
__result_use_check
Causes a warning to be emitted if a caller of the function
with this attribute does not use its return value. This is
known in gcc as "warn_unused_result" but we considered the
original naming unsuitable for an attribute.


git-svn-id: https://svn.freebsd.org/base/stable/10@280958 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f