o Make net.inet6.ip6.mif6table return special API structure, that doesn't   contain kernel pointers, and instead has interface index.   Bump __FreeBSD_version for that change. o Now, netstat/mroute6.c no longer needs to kvm_read(3) struct ifnet, and   no longer needs to include if_var.h
Note that this change is far from being a complete move of IPv6 multicast
routing to a proper API. Other structures are still dumped into their
sysctls as is, requiring userland application to #define _KERNEL when
including ip6_mroute.h and then call kvm_read(3) to gather all bits and
pieces. But fixing this is out of scope of the opaque ifnet project.

Sponsored by:	Nginx, Inc.
Sponsored by:	Netflix


git-svn-id: https://svn.freebsd.org/base/head@281172 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f