MFC r275539, r275554.
r275539:
cxgbe(4): Allow for different pad and pack boundaries for different
adapters.  Set the pack boundary for T5 cards to be the same as the
PCIe max payload size.  The chip likes it this way.

In this revision the driver allocate rx buffers that align on both
boundaries.  This is not a strict requirement and a followup commit
will switch the driver to a more relaxed allocation strategy.

r275554:
cxgbe(4): allow the driver to use rx buffers that do not end on a pack
boundary.


git-svn-id: https://svn.freebsd.org/base/stable/10@281212 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f